maxSize -> long

+++ src/RedCapped.Core/IMongoContext.cs
@@ -7,7 +7,7 @@ namespace RedCapped.Core
     public interface IMongoContext
     {
         Task<bool> CollectionExistsAsync(string collectionName);
-        Task CreateCappedCollectionAsync(string collectionName, int maxSize);
+        Task CreateCappedCollectionAsync(string collectionName, long maxSize);
         Task<IMongoCollection<BsonDocument>> GetCollectionAsync<T>(string collectionName, bool checkExists);
     }
 }
